requirements.txt
# UNet 目标识别项目依赖包

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 数据处理
numpy>=1.19.0
pillow>=8.0.0

# 可视化
matplotlib>=3.3.0
opencv-python>=4.5.0

# 工具库
tqdm>=4.60.0
tensorboard>=2.8.0

# 其他
scikit-learn>=0.24.0
